+++ app/src/main/java/com/breadwallet/tools/util/FileHelper.java
@@ -3,6 +3,11 @@ package com.breadwallet.tools.util;
 import android.util.Log;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.util.List;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipInputStream;
 
 /**
  * BreadWallet
@@ -76,5 +81,99 @@ public class FileHelper {
             sb.append("|  ");
         }
         return sb.toString();
+    }
+
+    public static void decompression(String srcPath, String outPath) throws Exception {
+        unZipFolder(srcPath, outPath);
+    }
+
+    public static void findAppJsonPath(File path, List<String> ret) {
+        if (null == path) return;
+        File[] files = path.listFiles();
+        for (File file : files) {
+            if (file.isDirectory()) {
+                findAppJsonPath(file, ret);
+            } else {
+                String name = file.getName();
+                if (name.equals("app.json")) {
+                    ret.add(file.getParent());
+                    return;
+                }
+            }
+        }
+    }
+
+    public static void unZipFolder(String zipFileString, String outPathString) throws Exception {
+        ZipInputStream inZip = new ZipInputStream(new FileInputStream(zipFileString));
+        ZipEntry zipEntry;
+        String szName = "";
+        while ((zipEntry = inZip.getNextEntry()) != null) {
+            szName = zipEntry.getName();
+            if (zipEntry.isDirectory()) {
+                szName = szName.substring(0, szName.length() - 1);
+                File folder = new File(outPathString + File.separator + szName);
+                folder.mkdirs();
+            } else {
+                File file = new File(outPathString + File.separator + szName);
+                if (!file.exists()) {
+                    file.getParentFile().mkdirs();
+                    file.createNewFile();
+                }
+                FileOutputStream out = new FileOutputStream(file);
+                int len;
+                byte[] buffer = new byte[1024];
+                while ((len = inZip.read(buffer)) != -1) {
+                    out.write(buffer, 0, len);
+                    out.flush();
+                }
+                out.close();
+            }
+        }
+        inZip.close();
+    }
+
+    public static String getJsonFromCapsule(File filePath) {
+        FileInputStream inputStream;
+        StringBuilder sb = new StringBuilder();
+        try {
+            inputStream = new FileInputStream(filePath);
+            byte buffer[] = new byte[1024];
+            int len;
+            while ((len = inputStream.read(buffer)) > 0) {
+                sb.append(new String(buffer, 0, len));
+            }
+            inputStream.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return sb.toString();
+    }
+
+    public static void deleteFile(File file) {
+        if (file.isDirectory()) {
+            File[] files = file.listFiles();
+            for (int i = 0; i < files.length; i++) {
+                File f = files[i];
+                deleteFile(f);
+            }
+            file.delete();
+        } else if (file.exists()) {
+            file.delete();
+        }
+    }
+
+
+
+    public static void deleteFileWithIgnore(File file, String ignore) {
+        if (file.isDirectory()) {
+            File[] files = file.listFiles();
+            for (int i = 0; i < files.length; i++) {
+                File f = files[i];
+                deleteFileWithIgnore(f, ignore);
+            }
+            file.delete();
+        } else if (file.exists() && !file.getName().equalsIgnoreCase(ignore)) {
+            file.delete();
+        }
     }
 }
