+++ app/src/main/res/drawable/ic_clear_cache.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="18dp"
+    android:viewportWidth="18"
+    android:viewportHeight="18">
+  <path
+      android:pathData="M9,9m-8.5,0a8.5,8.5 0,1 1,17 0a8.5,8.5 0,1 1,-17 0"
+      android:fillColor="#5A5465"/>
+  <path
+      android:pathData="M10.4821,8.1751L11.1892,7.468L13.0534,5.6038L12.7963,5.3467L12.5392,5.0896L10.675,6.9538L9.9679,7.6609L9.2608,6.9538L8.6179,6.311L7.718,5.411L7.1394,5.9896L12.1534,11.0036L12.732,10.425L11.8321,9.5251L11.1892,8.8823L10.4821,8.1751ZM11.4463,11.7107L6.4323,6.6967L4.6324,8.4966L5.4679,9.3321L6.0466,8.7534L6.7537,9.4605L6.175,10.0392L6.7537,10.6178L7.9751,9.3963L8.6822,10.1035L7.4608,11.3249L8.0394,11.9035L8.6179,11.325L9.325,12.0321L8.7465,12.6106L9.6465,13.5106L11.4463,11.7107ZM12.6034,8.8823L13.4391,9.7179C13.8296,10.1085 13.8296,10.7416 13.4391,11.1321L10.3536,14.2177C9.963,14.6082 9.3299,14.6082 8.9394,14.2177L3.9253,9.2037C3.5348,8.8131 3.5348,8.18 3.9253,7.7895L7.0109,4.7039C7.4014,4.3134 8.0346,4.3134 8.4251,4.7039L9.2608,5.5396L9.325,5.6038L9.9679,6.2467L10.675,5.5396L11.8321,4.3825C12.2226,3.992 12.8557,3.992 13.2463,4.3825L13.5034,4.6396L13.7605,4.8967C14.1511,5.2873 14.1511,5.9204 13.7605,6.311L12.6034,7.468L11.8963,8.1751L12.5392,8.818L12.6034,8.8823Z"
+      android:fillColor="#ffffff"
+      android:fillType="evenOdd"/>
+</vector>
