+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -1166,5 +1166,6 @@
     <string name="multi_wallet_already_current">已經是當前助記詞</string>
 
     <string name="send_memo_hint">Memo資訊會上鏈</string>
+    <string name="Clear_cache_title">清除緩存</string>
 
 </resources>
